Fix up the proxying a bit
We ended up with half open connection before. I attempted to
do a slightly nicer version of this with CloseRead, which
avoids the nasty check of the contents of an error string. But
that failed because the incoming conn is a TLSConn which doesnt
support ClosRead. FWIW, Kube do the same check